% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MATcor.R
\name{MATcor}
\alias{MATcor}
\title{MATcor: Matriz de correlacion grafica}
\usage{
MATcor(data, ...)
}
\arguments{
\item{data}{Dataframe con los datos.}

\item{...}{Variables especificas (sin comillas) a incluir en la matriz de correlacion.}
}
\value{
Devuelve (y ademas guarda en el entorno global) una lista con el grafico de correlaciones.
}
\description{
Esta funcion crea una matriz de correlacion grafica entre variables metricas de un dataframe.
Permite seleccionar un subconjunto especifico de variables.
}
